refactor to remove abstraction and keep on level of utility
