--- metrics/src/main/java/io/scalecube/metrics/api/Gauge.java
@@ -1,27 +1,0 @@
-package io.scalecube.metrics.api;
-
-/**
- * A gauge metric is an instantaneous reading of a particular value. To instrument a queue's depth, for example:<br>
- * 
- * <pre>
- * <code>
- * final Queue&lt;String&gt; queue = new ConcurrentLinkedQueue&lt;String&gt;();
- * final Gauge&lt;Integer&gt; queueDepth = new Gauge&lt;Integer&gt;() {
- * public Integer getValue() {
- * return queue.size();
- * }
- * };
- * </code>
- * </pre>
- *
- * @param <T> the type of the metric's value
- */
-public interface Gauge<T> {
-
-  /**
-   * Returns the metric's current value.
-   *
-   * @return the metric's current value
-   */
-  T getValue();
-}
