--- metrics/src/main/java/io/scalecube/metrics/api/GaugeFactory.java
@@ -1,9 +1,0 @@
-package io.scalecube.metrics.api;
-
-public interface GaugeFactory {
-
-  <T> Gauge<T> register(String component, String methodName, Gauge<T> gauge);
-
-  <T,G> Gauge<G> register(Class<T> component, String methodName, Gauge<G> gauge);
-
-}
