--- metrics/src/main/java/io/scalecube/metrics/api/Histogram.java
@@ -1,22 +1,0 @@
-package io.scalecube.metrics.api;
-
-/**
- * A metric which calculates the distribution of a value.
- */
-public interface Histogram {
-
-  /**
-   * Adds a recorded value.
-   *
-   * @param value the length of the value
-   */
-  public void update(int value);
-
-  /**
-   * Adds a recorded value.
-   *
-   * @param value the length of the value
-   */
-  public void update(long value);
-
-}
