+++ services/src/main/java/io/scalecube/services/Microservices.java
@@ -4,7 +4,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import io.scalecube.cluster.Cluster;
 import io.scalecube.cluster.ClusterConfig;
-import io.scalecube.metrics.api.MetricFactory;
+import io.scalecube.services.metrics.MetricFactory;
 import io.scalecube.services.routing.RoundRobinServiceRouter;
 import io.scalecube.services.routing.Router;
 import io.scalecube.transport.Address;
