+++ services/src/main/java/io/scalecube/services/ServiceProxyFactory.java
@@ -1,8 +1,7 @@
 package io.scalecube.services;
 
-import io.scalecube.metrics.api.Meter;
-import io.scalecube.metrics.api.MetricFactory;
-import io.scalecube.metrics.api.Metrics;
+import io.scalecube.services.metrics.MetricFactory;
+import io.scalecube.services.metrics.Metrics;
 import io.scalecube.services.routing.Router;
 import io.scalecube.transport.Message;
 
@@ -51,7 +50,7 @@ public class ServiceProxyFactory {
       @Override
       public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
 
-        mark(serviceInterface, metrics, method,"request");
+        Metrics.mark(serviceInterface, metrics, method, "request");
 
         Object data = method.getParameterCount() != 0 ? args[0] : null;
         final Message reqMsg = getRequestMessage(serviceDefinition, method, data);
@@ -64,14 +63,6 @@ public class ServiceProxyFactory {
         } else {
           return toReturnValue(method,
               dispatcher.invoke(reqMsg));
-        }
-
-      }
-
-      private <T> void mark(Class<T> serviceInterface, MetricFactory metrics, Method method, String eventType) {
-        if (metrics != null) {
-          Meter meter = metrics.meter().get(serviceInterface, method.getName(), eventType);
-          Metrics.mark(meter);
         }
       }
 
@@ -98,14 +89,14 @@ public class ServiceProxyFactory {
         } else if (method.getReturnType().equals(CompletableFuture.class)) {
           reuslt.whenComplete((value, ex) -> {
             if (ex == null) {
-              mark(serviceInterface, metrics, method,"response");
+              Metrics.mark(serviceInterface, metrics, method, "response");
               if (!Reflect.parameterizedReturnType(method).equals(Message.class)) {
                 future.complete(value.data());
               } else {
                 future.complete((T) value);
               }
             } else {
-              mark(serviceInterface, metrics, method,"error");
+              Metrics.mark(serviceInterface, metrics, method, "error");
               LOGGER.error("return value is exception: {}", ex);
               future.completeExceptionally(ex);
             }
