+++ services/src/main/java/io/scalecube/services/metrics/GaugeFactory.java
@@ -1,0 +1,11 @@
+package io.scalecube.services.metrics;
+
+import com.codahale.metrics.Gauge;
+
+public interface GaugeFactory {
+
+  <T> Gauge<T> register(String component, String methodName, Gauge<T> gauge);
+
+  <T, G> Gauge<G> register(Class<T> component, String methodName, Gauge<G> gauge);
+
+}
