+++ services/src/main/java/io/scalecube/services/metrics/MeterFactory.java
@@ -1,8 +1,11 @@
-package io.scalecube.metrics.api;
+package io.scalecube.services.metrics;
+
+import com.codahale.metrics.Meter;
 
 public interface MeterFactory {
 
   Meter get(String component, String methodName, String eventType);
 
   <T> Meter get(Class<T> component, String methodName, String eventType);
+
 }
