+++ services/src/main/java/io/scalecube/services/metrics/MetricFactory.java
@@ -1,0 +1,15 @@
+package io.scalecube.services.metrics;
+
+public interface MetricFactory {
+
+  MeterFactory meter();
+
+  TimerFactory timer();
+
+  CounterFactory counter();
+
+  GaugeFactory gauge();
+
+  HistogramFactory histogram();
+
+}
