+++ services/src/main/java/io/scalecube/services/metrics/TimerFactory.java
@@ -1,0 +1,11 @@
+package io.scalecube.services.metrics;
+
+import com.codahale.metrics.Timer;
+
+public interface TimerFactory {
+
+  Timer get(String component, String methodName);
+
+  <T> Timer get(Class<T> component, String methodName);
+
+}
