+++ services/src/test/java/io/scalecube/services/stress/GreetingServiceImpl.java
@@ -1,0 +1,19 @@
+package io.scalecube.services.stress;
+
+import io.scalecube.transport.Message;
+
+import java.util.concurrent.CompletableFuture;
+
+public class GreetingServiceImpl implements GreetingService {
+
+  @Override
+  public CompletableFuture<String> greeting(String name) {
+    return CompletableFuture.completedFuture("hello to: " + name);
+  }
+  
+  @Override
+  public CompletableFuture<Message> greetingMessage(Message request) {
+    return CompletableFuture.completedFuture(Message.fromData("hello to: " + request.data()));
+  }
+
+}
